Add F7Popup behavior and use it instead of subclass
