+++ src/app/scripts/views/home/popup.js
@@ -1,6 +1,13 @@
-var Popup = require('../popup');
+var Marionette = require('marionette');
+var F7Popup = require('../f7popup');
 
-module.exports = Popup.extend({
-    id: 'popup-home',
+module.exports = Marionette.View.extend({
+    behaviors: [
+      {
+        behaviorClass: F7Popup,
+        id: 'popup-home'
+      }
+    ],
+
     template: require('../../templates/home/popup.hbs')
 });
